deleted the extra NavLoggedIn component and updated the Nav since Sidebar already exists on large screen
